Update: module name and module samples
