+++ examples/webUtils/DojotHttpClient/package.json
