+++ lib/webUtils/DojotHttpClient.js
@@ -80,29 +80,32 @@ module.exports = class DojotClientHttp {
    * @param {*} reject The promise reject method of the previous request.
    * @param {*} previousConfigAndStatus The config of the previous request and retry status.
    *
+   * @return void
+   *
    * @private
    */
+  // eslint-disable-next-line consistent-return
   retry(
     requestError, options, resolve, reject, previousConfigAndStatus,
   ) {
     const outerThis = this;
     const { attempts, retryDelay, maxNumberAttempts } = previousConfigAndStatus;
+    this.logger.error(requestError.message);
     if (maxNumberAttempts > 0 && attempts >= maxNumberAttempts) {
-      reject(new Error('Number of attempts exceeded.'));
-    } else {
-      this.logger.error(requestError.message);
-      const newRetryDelay = (requestError.response && requestError.response.status === 429)
-        ? retryDelay * 2 : retryDelay;
-      this.logger.debug(`Retrying in ${retryDelay}`);
-
-      setTimeout(() => {
-        outerThis.logger.debug(`Retrying response - attempt:${attempts + 1}.`);
-        outerThis.doRequest(
-          options, resolve, reject, {
-            attempts: attempts + 1, retryDelay: newRetryDelay, maxNumberAttempts,
-          },
-        );
-      }, newRetryDelay);
+      return reject(new Error('Number of attempts exceeded.'));
     }
+
+    const newRetryDelay = (requestError.response && requestError.response.status === 429)
+      ? retryDelay * 2 : retryDelay;
+    this.logger.debug(`Retrying in ${retryDelay}`);
+
+    setTimeout(() => {
+      outerThis.logger.debug(`Retrying request - attempt:${attempts + 1}.`);
+      outerThis.doRequest(
+        options, resolve, reject, {
+          attempts: attempts + 1, retryDelay: newRetryDelay, maxNumberAttempts,
+        },
+      );
+    }, newRetryDelay);
   }
 };
