+++ lib/webUtils/framework/interceptors/keycloakAuthInterceptor.js
@@ -1,7 +1,7 @@
 const jwt = require('jsonwebtoken');
 const { TokenSet } = require('openid-client');
 const { Unauthorized } = require('../backing/errorTemplate');
-const DojotClientHttp = require('../../DojotClientHttp');
+const DojotHttpClient = require('../../DojotHttpClient');
 const KeycloakClientSession = require('../../KeycloakClientSession');
 
 
@@ -24,7 +24,7 @@ function formatCertificate(base64PublicKey) {
 async function verifyOnlineWithKeycloak(
   tokenRaw, tenant, configKeycloak, logger,
 ) {
-  const httpClient = new DojotClientHttp({
+  const httpClient = new DojotHttpClient({
     defaultClientOptions: {},
     logger,
     defaultRetryDelay: 5000,
