+++ lib/webUtils/index.js
@@ -2,7 +2,7 @@ const createServer = require('./createServer');
 const createTokenGen = require('./createTokenGen');
 const framework = require('./framework');
 const KeycloakClientSession = require('./KeycloakClientSession');
-const DojotClientHttp = require('./DojotClientHttp');
+const DojotHttpClient = require('./DojotHttpClient');
 const SecretFileHandler = require('./SecretFileHandler');
 
 module.exports = {
@@ -10,6 +10,6 @@ module.exports = {
   createTokenGen,
   framework,
   KeycloakClientSession,
-  DojotClientHttp,
+  DojotHttpClient,
   SecretFileHandler,
 };
